split funcs into own files to allow easier testing
